+++ src/app/common/file-validator.ts
@@ -19,4 +19,24 @@ export class FileValidator {
 
         return true;
     }
+
+    public isSupportedPlaylistFile(filePath: string): boolean {
+        const fileExtension: string = this.fileSystem.getFileExtension(filePath);
+
+        if (FileFormats.supportedPlaylistExtensions.includes(fileExtension.toLowerCase())) {
+            return true;
+        }
+
+        return false;
+    }
+
+    public isSupportedPlaylistImageFile(filePath: string): boolean {
+        const fileExtension: string = this.fileSystem.getFileExtension(filePath);
+
+        if (FileFormats.supportedPlaylistImageExtensions.includes(fileExtension.toLowerCase())) {
+            return true;
+        }
+
+        return false;
+    }
 }