+++ src/app/components/loading/loading.component.spec.ts
@@ -1,6 +1,6 @@
 import { IMock, Mock, Times } from 'typemoq';
 import { BaseDatabaseMigrator } from '../../common/data/base-database-migrator';
-import { BaseScheduler } from '../../common/scheduler/base-scheduler';
+import { BaseScheduler } from '../../common/scheduling/base-scheduler';
 import { BaseAppearanceService } from '../../services/appearance/base-appearance.service';
 import { BaseFileService } from '../../services/file/base-file.service';
 import { BaseIndexingService } from '../../services/indexing/base-indexing.service';