+++ src/app/services/playlist/playlist.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Observable, Subject } from 'rxjs';
-import { CollectionGrouper } from '../../common/collection-grouper';
+import { Collections } from '../../common/collections';
 import { FileValidator } from '../../common/file-validator';
 import { BaseFileSystem } from '../../common/io/base-file-system';
 import { Logger } from '../../common/logger';
@@ -48,6 +48,10 @@ export class PlaylistService implements BasePlaylistService {
 
     public get activePlaylistFolder(): PlaylistFolderModel {
         return this._activePlaylistFolder;
+    }
+
+    public get hasActivePlaylistFolder(): boolean {
+        return !this._activePlaylistFolder.isDefault;
     }
 
     public playlistsChanged$: Observable<void> = this.playlistsChanged.asObservable();
@@ -130,7 +134,7 @@ export class PlaylistService implements BasePlaylistService {
         }
 
         try {
-            const tracksToRemoveGroupedByPlaylistPath: Map<string, TrackModel[]> = CollectionGrouper.groupBy(
+            const tracksToRemoveGroupedByPlaylistPath: Map<string, TrackModel[]> = Collections.groupBy(
                 tracksToRemove,
                 (track) => track.playlistPath
             );